requirements.txt
django
nltk
matplotlib 
wordcloud
python-bidi